refactor: Move `ContainerHistoryLog` into `components` directory
